+++ src/components/AddButton/AddButton.module.scss
@@ -1,0 +1,30 @@
+@use '/src/styles/constants/colors';
+@use '/src/styles/mixins/position';
+
+.addButton {
+  @include position.absolute(auto 40px 40px auto);
+  width: 66px;
+  height: 66px;
+  background-color: colors.$BLUE;
+  border-radius: 50%;
+  box-shadow: 0 10px 20px 0 rgba(colors.$BLACK, 15%);
+  transition: 0.2s;
+
+  &:hover {
+    background-color: rgba(colors.$BLUE, 90%);
+  }
+
+  &::before,
+  &::after {
+    @include position.middleBox;
+    width: 20px;
+    height: 2px;
+    content: '';
+    background-color: colors.$WHITE;
+  }
+
+  &::after {
+    transform: translate(-50%, -50%) rotate(90deg);
+    transform-origin: center;
+  }
+}
