+++ src/components/SearchBar/styles.module.scss
@@ -15,7 +15,7 @@
   align-items: center;
   height: 30px;
   padding: 0.7rem 1rem;
-  background-color: colors.$GRAYE;
+  border: 1px solid colors.$GRAYE;
   border-radius: 5px;
 }
