+++ src/components/SmallButton/smallButton.module.scss
@@ -8,7 +8,7 @@
   height: 30px;
   background-color: colors.$WHITE;
   border: 1px solid colors.$GRAYD;
-  fill: colors.$GRAY2;
   border-radius: 50%;
+  fill: colors.$GRAY2;
   transition: 0.2s;
 }
