+++ src/components/TodoFilter/TodoFilter.module.scss
@@ -1,0 +1,25 @@
+@use '/src/styles/constants/colors';
+@use '/src/styles/mixins/flexbox';
+
+.toggleInputWrapper {
+  @include flexbox.flexbox(center center);
+  justify-content: space-around;
+  padding-bottom: 10px;
+  color: colors.$TASKS_TITLE;
+}
+
+.toggleTab {
+  width: 100px;
+  padding: 4px;
+  text-align: center;
+  border-radius: 3px;
+
+  &:hover {
+    color: colors.$BLUE;
+  }
+
+  &.activeTab {
+    font-weight: 600;
+    color: colors.$BLUE;
+  }
+}
