+++ src/components/TodoList/TodoList.jsx
@@ -1,12 +1,12 @@
 import PropTypes from 'prop-types'
-import TodoListItem from '../TodoListItem/TodoListItem'
+import TodoItem from '../TodoItem/TodoItem'
 import styles from './TodoList.module.scss'
 
 function TodoList({ todoList, setTodoList }) {
   return (
     <ul className={styles.tasks}>
       {todoList.map((todo) => (
-        <TodoListItem key={todo.id} todo={todo} setTodoList={setTodoList} />
+        <TodoItem key={todo.id} todo={todo} setTodoList={setTodoList} />
       ))}
     </ul>
   )
