+++ src/components/TodoList/TodoList.module.scss
@@ -15,12 +15,6 @@
     height: 100%;
   }
 
-  h1 {
-    font-size: 32px;
-    font-weight: 600;
-    color: colors.$BIG_TITLE;
-  }
-
   .tasksTitle {
     margin: 30px 0 24px;
     font-size: 12px;
@@ -53,14 +47,6 @@
     }
   }
 
-  .tasks {
-    display: flex;
-    flex-direction: column;
-    height: 75%;
-    margin: 2px;
-    overflow: auto;
-  }
-
   .addButton {
     @include position.absolute(auto 40px 40px auto);
     width: 66px;
@@ -84,16 +70,11 @@
     }
   }
 }
-// h1 {
-//   font-size: 32px;
-//   font-weight: 600;
-//   color: colors.$BIG_TITLE;
-// }
 
-// .tasks {
-//   display: flex;
-//   flex-direction: column;
-//   gap: 10px;
-//   max-height: 100%;
-//   overflow: scroll;
-// }
+.tasks {
+  display: flex;
+  flex-direction: column;
+  height: 70%;
+  margin: 2px;
+  overflow: auto;
+}
