--- src/components/TodoListItem/TodoListItem.module.scss
@@ -1,58 +1,0 @@
-@use '/src/styles/constants/colors';
-@use '/src/styles/mixins/flexbox';
-@use '/src/styles/mixins/position';
-@use '/src/styles/constants/levels';
-
-.task {
-  display: flex;
-  align-items: center;
-  height: 72px;
-  line-height: 72px;
-  background-color: colors.$WHITE;
-  border-radius: 32px;
-  box-shadow: 0 10px 20px 0 rgba(colors.$BLACK, 2%);
-
-  .checkboxWrapper {
-    @include flexbox.flexbox(center center);
-    position: relative;
-    width: 28px;
-    height: 28px;
-    margin: 0 24px;
-
-    svg {
-      @include position.middleBox;
-      z-index: levels.$CHECK_ICON;
-      color: colors.$WHITE;
-      pointer-events: none;
-      opacity: 0;
-      transition: 0.2s;
-    }
-  }
-
-  input[type='checkbox'] {
-    @include position.middleBox;
-    z-index: levels.$CHECK_BOX;
-    width: 28px;
-    height: 28px;
-    cursor: pointer;
-    border: 2px solid colors.$BLUE;
-    border-radius: 50%;
-    transition: 0.2s;
-
-    &:checked {
-      background-color: colors.$BLUE;
-      border: 0;
-
-      ~ svg {
-        opacity: 1;
-      }
-    }
-  }
-
-  .title {
-    overflow: hidden;
-    color: colors.$TASK_ITEM_TITLE;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-}
