+++ src/routes/home/Home.module.scss
@@ -1,0 +1,16 @@
+@use '/src/styles/constants/colors';
+
+.tasksTitle {
+  margin: 30px 0 24px;
+  font-size: 12px;
+  font-weight: 600;
+  color: colors.$TASKS_TITLE;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+}
+
+h1 {
+  font-size: 32px;
+  font-weight: 600;
+  color: colors.$BIG_TITLE;
+}
